fix: Optimize admin auth login and verification response
