+++ apps/backend/src/controllers/admin/auth.ts
@@ -8,6 +8,8 @@ import {
 } from "@/types/admin/auth";
 import { IForgetPasswordWithEmail } from "@/types/customer";
 import {
+  generateAccessToken,
+  generateRefreshToken,
   sendError,
   verifyField,
   cookieOptions,
@@ -19,7 +21,6 @@ import {
   generateSmsCode,
   createPassword,
   generateStageToken,
-  issueAdminTokens,
 } from "@/utils";
 import { NextFunction, Response, Request } from "express";
 
@@ -58,14 +59,29 @@ export const login = async (
     }
 
     if (!user.twoFactorAuth) {
-      logInfo("Admin logged in without 2FA", {
-        ip: (req as any).hashedIp,
-        userId: user.id,
-        event: "admin_login_success",
+      const payload = { id: user.id, email: user.email };
+      const access = generateAccessToken(payload, "ADMIN");
+      const refresh = generateRefreshToken(payload, "ADMIN");
+
+      res.cookie("accessToken", access.token, {
+        ...cookieOptions,
+        maxAge: remember ? refresh.expiresIn : undefined,
+      });
+      res.cookie("refreshToken", refresh.token, {
+        ...cookieOptions,
+        maxAge: remember ? refresh.expiresIn : undefined,
       });
 
-      const result = issueAdminTokens(res, user, remember);
-      return res.status(200).json(result);
+      logInfo("Admin logged in successfully", {
+        ip: (req as any).hashedIp,
+        userId: user.id,
+        event: "admin_verified",
+      });
+
+      return res.json({
+        message: getResponseMessage("loginSuccessful"),
+        data: { user: { email: user.email } },
+      });
     }
 
     const {
@@ -95,6 +111,7 @@ export const login = async (
       ...cookieOptions,
       maxAge: 5 * 60 * 1000,
     });
+
     return res.status(200).json({ message: getResponseMessage("codeSent") });
   } catch (error: unknown) {
     logCatchyError("Login exception", error, {
@@ -155,7 +172,19 @@ export const verifyOtp = async (
     res.clearCookie("admin_verify_stage");
 
     const remember = (req as any).remember;
-    const result = issueAdminTokens(res, admin, remember);
+    const payload = { id: admin.id, email: admin.email };
+    const access = generateAccessToken(payload, "ADMIN");
+    const refresh = generateRefreshToken(payload, "ADMIN");
+
+    res.cookie("accessToken", access.token, {
+      ...cookieOptions,
+      maxAge: remember ? refresh.expiresIn : undefined,
+    });
+
+    res.cookie("refreshToken", refresh.token, {
+      ...cookieOptions,
+      maxAge: remember ? refresh.expiresIn : undefined,
+    });
 
     logInfo("Admin verified successfully", {
       ip: (req as any).hashedIp,
@@ -163,7 +192,10 @@ export const verifyOtp = async (
       event: "admin_verified",
     });
 
-    return res.json(result);
+    return res.json({
+      message: getResponseMessage("loginSuccessful"),
+      data: { user: { email: admin.email } },
+    });
   } catch (error) {
     logCatchyError("Admin verify otp exception", error, {
       ip: (req as any).hashedIp,
