+++ apps/backend/src/utils/auth/index.ts
@@ -1,2 +1,1 @@
 export * from "./isAuthenticated";
-export * from "./issueAdminTokens";
